requirements.txt
numpy==1.14.5
matplotlib==2.2.2
scipy==0.19.1
permute==0.1a4
scikit_learn==0.19.2
PyYAML==3.13
lightgbm>=3.0.0
